Ensured the session management and fixed the getEndpoint method to fetch the endpoints
