+++ index.js
@@ -1,4 +1,4 @@
-import apiService from './apiService.js';
+import apiService from './src/apiService.js';
 
 // Exported the apiService for external use
 export default apiService;
