+++ src/apiConfig.js
@@ -1,15 +1,3 @@
-export function getApiConfig(key) {
-  const config = {
-    API_URL: process.env.API_URL, 
-    API_USERNAME: process.env.API_USERNAME, 
-    API_PASSWORD: process.env.API_PASSWORD, 
-    PRACTICE: process.env.PRACTICE,
-  };
-
-  return config[key];
-}
-
-
 export const endpoints = {
     abbreviations : 'db/abbreviations',
     acceptable_ips : 'db/acceptable_ips',
